Add support for TINK cryptography library
This brings in the TINK library. TINK is q multi-language, cross-platform
library that provides cryptographic APIs that are secure, easy to use
correctly, and hard(er) to misuse.

We need the TINK library to validate JWT tokens to protect the gRPC
endpoint.

This contains a set of test vectors from google wycheproof. See
https://github.com/google/wycheproof for details.

Includes a series of windows fixes.

+++ android/build/python/aemu/cmake.py
@@ -48,7 +48,10 @@ def configure(args, target):
     if args.clean:
         if os.path.exists(args.out):
             logging.info("Clearing out %s", args.out)
-            shutil.rmtree(args.out)
+            if platform.system() == "Windows":
+                run(['rmdir','/S', '/Q', args.out])
+            else:
+                shutil.rmtree(args.out)
         if not os.path.exists(args.out):
             os.makedirs(args.out)
